Work around faulty Xbox emulation software.
SCP Toolkit caused a crash where it would send messages to libusb's
callback. Since these weren't expected, a pointer would be wrong, and
a hard crash would occur.

This change looks for an Xbox 360 controller, and disables HIDAPI
if it sees one, as it's not the pro controller we care about.