Fix a problem where the fullscreen toggle wasn't present on macOS.
This looks like an issue with SDL2 that needs to be addressed
upstream at some point. The issue is that once the application
was toggled to fulscreen, it would flag itself as non-resizeable,
which would remove the toggle button and menu item (at leas on
Big Sur).

Fixes renpy/renpy#2623.