Rewrite of how the Android splashscreen works.
Instead of trying to use SDL's non-GL mode to display the presplash,
display it on a layer at startup, then have Ren'Py hide that
layer after the first frame.

This fixes a lot of Android weirdness, where the change in
surface would cause problems. Most notably, black screens
as the game starts.