A new approach to iOS builds.
This uses llvm directly, without the need for cctools-port. This
makes it possible to choose if we're compiling for ios or the
ios simulator.

While this is enough to get Ren'Py to start on the iOS simulator,
there is a crash on dissolve. I'm documenting that for now and
focusing on getting 7.4.7 out the door.