Android .gz handling counter-measures
AAPT unavoidably gunzips files with a .gz extension as part of it's
packaging step, decompressing their contents and dropping the .gz from
their file name.

While it's true that zipping an already compressed file isn't the most
ideal thing to do, unfortunately the result for creators is a missing
file due to the name change, and a need to know about this quirk as well
as introduce an Android specific code path to handle it.

In order to avoid this scenario, prior to the packaging step we
temporarily double gzip affected files, then leave AAPT to unpack them
back into the original location.