web3 pwa: Redo how offline caching works.
Rather than trying to load and sync the cache, we prefer fetch,
which will use the cache if it can and the cache is up to date
with the server.

Only if the fetch fails does the service worker look in its
cache, and try to fulfill the request that way.