rapt: Create a single keyfile for apk and bundle.
This is the outcome of https://lemmasoft.renai.us/forums/viewtopic.php?f=8&t=66547 .

The distinction between bundle.keystore and android.keystore was always
pretty arbitrary - it was a way to work around an issue with older
versions of the play store. That doesn't seem to be a problem,
especially now that the play store is requiring aab uploads,
and so this can go away.

Some now-unnecessary prompts were removed, rather than changed,
so that translations don't need to be updated.