Defer pyc compilation until pythonlib task
The pythonlib task is already responsible for assembling the
stdlib/site-packages amalgam that we ship with Ren'Py, so rather than
scattering our pyc preferences throughout various tasks, instead
force-recompile them just before copying them in to the build output.

As an extension to this, the Context class gets a new compile method for
compiling py files. It encapsulates our preferences based on the version
of Python for which we are building, and allows for centralised filename
elision. This not only allows for more succinct error reporting, but
also prevents build server file paths being baked into the output.