Add renpy-specific fullscreen API.
This fixes issues with fullscreen, especially on mobile platforms
where the device can rotate while fullscreen is enabled. At the
same time, Ren'Py has been handling web-resize well.

So, we simply requestFullscreen on the html element, and let it
bring the window fullscreen, resizing the canvas as it does so.