rapt: Android changes.
This combines two changes. The first is to move to using onStop
for the shutdown process, as onStop is not called when the
app is visible but not foreground. This allows the sound to
keep running when an app is overlaying the window. (Like a
translator app.)

Fixes renpy/renpy#5448.

This also fixes an issue where onPause/onStop could be called
without the app foregrounding (My guess is that it was
happening when the live display kicked in, but no idea why.)
This meant onPause (now onStop) would wait forever for finshOnPause,
which would never be called.

There's now a new armOnStop that needs to be called at the start
of execution and when starting up after a stop. If not called,
onStop completes quickly.