Periodically check if the persistent data needs to be stored
The persistent data is generally flushed to disk when Ren'Py exits. This
is not possible on Web, because no code can be executed after the
"beforeunload" event that the browser sends before leaving the page. IIRC
this event is handled correctly by emscripten, but this makes sdl2 and/or
Ren'Py post another event in their internal events queue which is never
processed then.

The current workaround for Web is to flush the persistent data each time
the user performs a manual save. This is better than nothing, but that's
pretty frustrating when you change some preferences in a game, then start
playing it, decide to reload the page and realize all your preferences
have been reset because the persistent data has not been stored to disk.

This commit periodically checks for changes in the persistent data and
stores it to the Web storage if so. The current period is 5 seconds, and
it takes ~50 ms on my PC to check for that, so it should not impact the
performances too much?