Move to LLVM 18.
LLVM 19 wasn't working due to a known and fixed issue with libffi.
Once that comes out, we should be able to move up.

There is an issue linking with libGL, but that's dynamically loaded,
so no big deal.

Zlib needed to be upgraded to compile with the new compiler, now it
works.