Restore file mtime when extracting game.zip
The zipfile lib does not restore the files modification time when
extracting a ZIP file. This has an impact on Python as the PY file
mtime is stored in the PYC header so that Python can check if the
PYC file needs to be recompiled before loading it: if the mtime
for the PY file does not match the value in the PYC header, the file
is recompiled. So on Web, all the PY files are recompiled, which
increases the game load time by at least 10% on my tests.

This commit update the file mtime after extraction using the value
in the ZIP entry to prevent that issue.